Update 127_0_0_1.sql
This is the exported Database. 
The python script needs a db named 'myarduinodb' and a table named 'arduino'. The php code needs a table named average
If you dont want to import this db you have to create one as in line 19 and two tables as in line 28 and line 55
THIS DB HAS ALREADY SOME VALUES INSERTED (so the site isn't blank)